Move ubuntu bastion's user-data/cloud-init to the launch template
The user data block has been moved to the launch template.

The cloud-init.sh isn't called by the instance's user data script
anymore, but it is pulled in and executed via a cfn-init command.

This allows us to safely inject data from the template without
needing to care much or at all what data that is, how to properly quote
it in the context of a script -- we can expose these now to the
environment of the command directly.